Start of more datafile writing and reading options
I need to be able to write the trees themselfes to a separate datafile
if I want to read them back in for the differences between and inside
trees.

I haven't tested it, because I need to clock out and it's Friday. This
is a bad GitHub practice, but I'm working on my own so it's more okay.